requirements.txt
aiogram==3.7.0
aiosqlite==0.20.0
anthropic==0.34.0
telethon==1.36.0
matplotlib==3.9.0
aiohttp==3.10.0
python-dotenv==1.0.1
